feat(core): add token utils module and tests
