+++ .changeset/token-utils-refactor.md
@@ -1,0 +1,5 @@
+---
+'@lapidist/design-lint': patch
+---
+
+refactor token utilities into shared module
