+++ src/rules/token-animation.ts
@@ -4,7 +4,7 @@ import {
   matchToken,
   extractVarName,
   closestToken,
-} from '../utils/token-match.js';
+} from '../core/token-utils.js';
 
 export const animationRule: RuleModule = {
   name: 'design-token/animation',
