+++ src/rules/token-border-color.ts
@@ -6,7 +6,7 @@ import {
   matchToken,
   extractVarName,
   closestToken,
-} from '../utils/token-match.js';
+} from '../core/token-utils.js';
 
 type ColorFormat =
   | 'hex'
