+++ src/rules/token-box-shadow.ts
@@ -4,7 +4,7 @@ import {
   matchToken,
   extractVarName,
   closestToken,
-} from '../utils/token-match.js';
+} from '../core/token-utils.js';
 
 export const boxShadowRule: RuleModule = {
   name: 'design-token/box-shadow',
