+++ src/rules/token-opacity.ts
@@ -5,7 +5,7 @@ import {
   matchToken,
   extractVarName,
   closestToken,
-} from '../utils/token-match.js';
+} from '../core/token-utils.js';
 import { isStyleValue } from '../utils/style.js';
 
 export const opacityRule: RuleModule = {
