--- tests/token-match.test.ts
@@ -1,30 +1,0 @@
-import test from 'node:test';
-import assert from 'node:assert/strict';
-import {
-  matchToken,
-  closestToken,
-  extractVarName,
-} from '../src/utils/token-match.ts';
-
-void test('matchToken handles regexp and glob patterns and missing matches', () => {
-  assert.equal(matchToken('--brand-primary', [/^--brand-/]), '--brand-primary');
-  assert.equal(matchToken('--brand-primary', ['--brand-*']), '--brand-primary');
-  assert.equal(matchToken('--foo', ['--bar']), null);
-});
-
-void test('matchToken is case-insensitive for string patterns', () => {
-  assert.equal(matchToken('--BRAND-primary', ['--brand-*']), '--BRAND-primary');
-});
-
-void test('closestToken skips non-string patterns and suggests best match', () => {
-  assert.equal(closestToken('--baz', ['--bar', /^--foo-/]), '--bar');
-  assert.equal(closestToken('--baz', [/^--foo-/]), null);
-});
-
-void test('extractVarName parses var() and ignores invalid values', () => {
-  assert.equal(extractVarName('var(--x)'), '--x');
-  assert.equal(extractVarName('var(--x, 10px)'), '--x');
-  assert.equal(extractVarName('var(  --x  )'), '--x');
-  assert.equal(extractVarName('--x'), null);
-  assert.equal(extractVarName('var(--foo.bar)'), null);
-});
